+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authz/responses/TupleDeleteResponse.java
@@ -1,0 +1,6 @@
+package cloud.pangeacyber.pangea.authz.responses;
+
+import cloud.pangeacyber.pangea.Response;
+import cloud.pangeacyber.pangea.authz.results.TupleDeleteResult;
+
+public final class TupleDeleteResponse extends Response<TupleDeleteResult> {}